+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - an api (`knwler.collect`) to fetch url as markdown, fetch documents (pdf, xlsx...) and Wikipedia article. All cached for convenience.
 - you can now give a URL to a pdf or just a webpage, all the rest happens automatically for you
 - a higher level API (`knwler.api`) has been added to simplify downstream integrations
+- the CLI has a new `fetch` command allowing to fetch data and parse in one go
 
 ## v0.4.1